create file-picker store, reorganize files
