+++ frontend/src/store/files/file-picker-store.ts
@@ -1,0 +1,13 @@
+import { Store } from '@tanstack/react-store'
+
+type FilePickerState = {
+  isOpen: boolean
+}
+
+export const filePickerStore = new Store<FilePickerState>({ isOpen: false })
+
+export const filePickerActions = {
+  toggleDialog: () => {
+    filePickerStore.setState((prev) => ({ isOpen: !prev.isOpen }))
+  },
+}
