+++ frontend/src/store/files/files-store.ts
@@ -1,0 +1,1 @@
+type FilesState = {}
